requirements.txt
requests
beautifulsoup4
feedparser
python-dotenv
cachetools
praw